Improve the handling of forwarding Start/Stop
